Notification optimization and quick post frontend

+++ server/services/noteService.ts
@@ -27,19 +27,19 @@ export async function addNote(noteData: INoteDB) {
 */
 export async function deleteNote({studentDocID, noteDocID}: IManageUserNote) {
     try {
-        await Notes.deleteOne({ _id: noteDocID })
-        await Students.updateOne(
-            { _id: studentDocID },
-            { $pull: { owned_notes: noteDocID } }
-        )
-        await Comments.deleteMany({ _id: noteDocID })
-    
-        let noteNotifs = await Notifs.find({ docType: { $in: ["note-feedback", "note-reply", "note-mention"] } })
-        let noteSpecificNotifsDocIDs = noteNotifs.filter(noti => noti["noteDocID"].toString() === noteDocID).map(noti => noti["_id"])
-        await Notifs.deleteMany({ _id: { $in: noteSpecificNotifsDocIDs } })
-    
-        await deleteNoteImages({ studentDocID, noteDocID })
-
+        let deleteResult = await Notes.deleteOne({ _id: noteDocID })
+        console.log(deleteResult)
+        if (deleteResult.deletedCount !== 0) {
+            await Students.updateOne(
+                { _id: studentDocID },
+                { $pull: { owned_notes: noteDocID } }
+            )
+            await Comments.deleteMany({ _id: noteDocID })
+            let noteNotifs = await Notifs.find({ docType: { $in: ["note-feedback", "note-reply", "note-mention"] } })
+            let noteSpecificNotifsDocIDs = noteNotifs.filter(noti => noti["noteDocID"].toString() === noteDocID).map(noti => noti["_id"])
+            await Notifs.deleteMany({ _id: { $in: noteSpecificNotifsDocIDs } })
+            await deleteNoteImages({ studentDocID, noteDocID })
+        }
         return true
     } catch (error) {
         return false
@@ -117,7 +117,7 @@ export async function getNoteForShare({noteDocID, studentDocID}: IManageUserNote
 }
 
 export async function getAllNotes(studentDocID: string, options = { skip: 0, limit: 3 }) {
-    let notes = await Notes.find({}, { ownerDocID: 1, title: 1, content: 1, feedbackCount: 1, upvoteCount: 1 })
+    let notes = await Notes.find({ completed: true }, { ownerDocID: 1, title: 1, content: 1, feedbackCount: 1, upvoteCount: 1 })
         .sort({ createdAt: -1 })
         .skip(options.skip)
         .limit(options.limit)
